backend/requirements.txt
# Backend dependencies for StochastiQdata
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
supabase==2.3.4
pyjwt[crypto]==2.8.0
httpx>=0.24,<0.26
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
